added support us screen android variation
